Downloader: rename queue client module
