+++ docs/investigations/2026-01-03_PROD_SPEED_PARITY_YAPPI.md
@@ -29,7 +29,7 @@ From `codexspx3280b3dee4b7410fba_profile_yappi_raw.csv`:
 - `/usr/local/lib/python3.12/threading.py Condition.wait` (`ttot_s ≈ 6003`)
 - `/usr/local/lib/python3.12/threading.py Event.wait` (`ttot_s ≈ 4005`)
 - `/usr/local/lib/python3.12/queue.py Queue.get` (`ttot_s ≈ 1999`)
-- `lumibot.tools.thetadata_queue_client.QueueClient.execute_request / wait_for_result` (`ttot_s ≈ 1048 / 1029`)
+- `lumibot.tools.data_downloader_queue_client.QueueClient.execute_request / wait_for_result` (`ttot_s ≈ 1048 / 1029`)
 - `lumibot.tools.thetadata_helper.get_request / get_historical_data` (`ttot_s ≈ 1050 / 1289`)
 - `lumibot.backtesting.thetadata_backtesting_pandas.ThetaDataBacktestingPandas.get_quote` (`ttot_s ≈ 1289`)
 
@@ -54,7 +54,7 @@ Key difference observed in logs:
 1. **Deploy v4.4.22** (includes the index fast-path perf fix) to production.
 2. Re-run the same profiled production backtest and compare:
    - wall time
-   - `ncall` for `thetadata_queue_client.queue_request`, `thetadata_backtesting_pandas.get_quote`
+   - `ncall` for `data_downloader_queue_client.queue_request`, `thetadata_backtesting_pandas.get_quote`
    - share of time in `threading/queue` waits vs non-waiting work
 3. If production is still materially slower after request-count reduction:
    - investigate downloader throughput / queue contention
